lint and a bunch of stuff
