+++ weave/autopatch.py
@@ -23,3 +23,7 @@ def autopatch_openai() -> None:
 
 def autopatch() -> None:
     autopatch_openai()
+
+    from .trace.integrations.mistral.mistral import autopatch_mistral
+
+    autopatch_mistral()
