+++ weave/trace/integrations/mistral/mistral_test.py
@@ -4,25 +4,30 @@ import weave
 from weave.trace_server import trace_server_interface as tsi
 from mistralai.client import MistralClient
 from mistralai.models.chat_completion import ChatMessage
+from .mistral import autopatch_mistral, undo_patch_mistral
 
+from typing import Generator
 
 # Add to some docs, to re-record, and run:
 # `MISTRAL_API_KEY=... pytest --weave-server=prod --record-mode=rewrite trace/integrations/mistral/mistral_test.py::test_mistral_quickstart`
 
 
+@pytest.fixture(scope="module")
+def patch_mistral() -> Generator[None, None, None]:
+    autopatch_mistral()
+    yield
+    undo_patch_mistral()
+
+
 @pytest.mark.vcr(
     filter_headers=["authorization"], allowed_hosts=["api.wandb.ai", "localhost"]
 )
-def test_mistral_quickstart(client: weave.weave_client.WeaveClient) -> None:
+def test_mistral_quickstart(
+    client: weave.weave_client.WeaveClient, patch_mistral: None
+) -> None:
     # This is taken directly from https://docs.mistral.ai/getting-started/quickstart/
     api_key = os.environ.get("MISTRAL_API_KEY", "DUMMY_API_KEY")
     model = "mistral-large-latest"
-
-    # move this to init somewhere
-    from .mistral import mistral_patches
-
-    for patch in mistral_patches.values():
-        patch.attempt_patch()
 
     mistral_client = MistralClient(api_key=api_key)
